Today I learned many things about Facebook's quote APIs quote, but mostly I learned that I want to grab a member of the Facebook API team and choke them.
There are a broad array of acceptable responses to configuration errors. They
include:

- providing a return value from your function which indicates a problem
- throwing an exception when the lack of configuration is discovered
- using your logging system to report the configuration problem
- calling the developer's mother and telling her that he's been a bad boy

What is not acceptable is silently failing. I don't care how deep down in the
bowels of your system you are when you discover that I don't happen to have a
Facebook Connect URL configured in my application. If XFBML is going to fail to
work because this totally unrelated configuration parameter of a totally
unrelated feature of your API is not set, you should goddamned well report that
somehow. Send a fucking carrier pigeon if you have to.


git-svn-id: svn+ssh://src.earth.threerings.net/everything/trunk@187 a9e240f3-4568-de11-90a4-003048321218