Further refinements to grid generation. Now:
1. If you choose a Gimme an R powerup, we attempt to give you a rarity R card
that you need (ie. from a set you're collecting and that you don't already
have). If you choose any other powerup or none at all, we select a bonus card
(rarity V and up) that you need.

2. We then select 3 (or 4 if we were unable to give you a card you need in step
1) cards that you need. This is scaled down if you're collecting fewer than 3
(or 4) sets.

3. We then select up to 4 more cards (again scaled down for small collections)
from categories you are collecting, but don't care if they're duplicates or not
(this tends to result in giftable cards as you and your friends are likely to
overlap somewhat in your collections). This step is skipped if you select
ALL_NEW_CARDS. We fill your grid in this step if you choose NO_SURPRISES.

4. We then fill out the remaining slots in your grid with random cards,
honoring ALL_NEW_CARDS if that powerup was used.

I think this will increase the rate of grid completion. I plan to continue to
tweak grid generation until players average one completed series per week.


git-svn-id: svn+ssh://src.earth.threerings.net/everything/trunk@510 a9e240f3-4568-de11-90a4-003048321218