I was going to do something complicated, but then I remembered, simple is good.
We now support seed things by just tacking &seed=XXX onto the landing URL. The
ID of every series is shown now on the series editing page, so it can be easily
looked up for any desired series. A rarity I card is chosen from the seed
series and granted to the player when their data is first created. Simple!


git-svn-id: svn+ssh://src.earth.threerings.net/everything/trunk@614 a9e240f3-4568-de11-90a4-003048321218