Rejigger! - Don't let users set a "neutral" preference: it's gotta be like or dislike. - Rejiggered the way likes are sent to the client once I discovered that the thingIds on the grid are without categories. I would have had to include in the grid a mapping of thingId -> category and then another of the likes for categories. And only for things already flipped, otherwise clients could hack and gain info on their grid. And the grid gets resolved on every flip, so I would have either redone this extra work each time or complicated some other stuff...
So I changed things so that your like/dislike prefs are sent down when
you log in, and used for the duration of your session. This actually
simplified a few things, because now they come out of the Context
instead of needing to be passed everywhere. Instead I make the decision
where needed about whether we show the LikeWidget or not.
And if we change our mind and allow you to configure your likes while
viewing someone else's collection it's a simple few lines of code to do so.


git-svn-id: svn+ssh://src.earth.threerings.net/everything/trunk@752 a9e240f3-4568-de11-90a4-003048321218