Previously, we created friend mappings for all of your facebook friends that had samsara accounts.
I'm not sure why.
We then had to specifically filter out the folks that didn't
have Everything accounts- either explicitly or implicitly when
loading records of card holdings.

I think it'll be better to just track friendships of people who actually
have everything accounts, and reduce the set of ids passed to the database
in various places. There is possibly a way to do this by using
friends_getAppUsers rather than friends_get, but the googley API we're
using doesn't provide a nice interface to that function, so we'll
just load up their names and see who's missing.


git-svn-id: svn+ssh://src.earth.threerings.net/everything/trunk@769 a9e240f3-4568-de11-90a4-003048321218