Fix the non-panic-worthy issue with picking things with a very large exclude list- even if we make a filtered copy of the ThingList...
If we spin in circles trying to pick the things randomly, recurse with
a forced filtered copy of the ThingList, removing the things already picked.
This should always work- eventually you'll either get enough things
or you'll clear out the list of eligible things. I lowered the number of
'miss' iterations before doing the forced copy to 256. I cut back on
the logging too, because I think that the only time we should see this now
is when the "almost everything" players use ANIGN.


git-svn-id: svn+ssh://src.earth.threerings.net/everything/trunk@800 a9e240f3-4568-de11-90a4-003048321218