Added notice system, played-mobile reward, some movery.
The notices will be displayed by the client, and chiefly serve to inform you
that you got some coins for some fine reason.

The first notice is that you get coins for playing via the mobile client (iff
you've already played the non-mobile client).

Said notices are not yet wired up on the client, that's next. I also plan to
implement a notice+reward when a new friend starts playing. We may not be able
to automate the process of spamming your friends, but we can reward you for it
and leave you to devise your own mechanisms.