Disable most of the game.
This disables access to the game and forces the news to be displayed.

We lost access to all the Everything media and I have no way to get it back.

Alas that's the end of the line for The Everything Game. It was a good run. Not
quite ten years, but we'll round up since 9 is pretty close to 10. RIP.